requirements.txt
asn1crypto==0.24.0
boto==2.49.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
Django==2.2.1
django-datatables-view==1.17.0
django-model-utils==3.1.2
django-mysql==3.0.0.post1
django-notifications-hq==1.5.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.3
html2text==2018.1.9
http-ece==1.1.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
pkg-resources==0.0.0
py-vapid==1.5.0
pycparser==2.19
pyfcm==1.4.5
python-resize-image==1.1.19
pytz==2019.1
pywebpush==1.9.4
qrcode==6.1
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
uWSGI==2.0.18
